Separe los juegos y agregue un juego para adivinar la palabra, y en el main solo se encuentra el menu
