+++ nombres.txt
@@ -1,1 +1,3 @@
 Nombre: Eduardo
+Nombre: Miriam
+
